Replace a limitation by another one
Limit 2. "_<html> - <head> - <body> structure is mandatory_" has been
replaced by "_strict self-closing tags constraint (`/>` and not `>`)_".

This comes from the way Pasap sees the code which it has to parse. Until
now, it has been seeing it as HTML. It has allowed compatibility with any
HTML5 code.

Now, the paresr will see the input code as XML. This is more strict about
self-closing tags, but it will be simpler to implement namespaces for the
custom tags.